rename lib file
